flag graphql queries moved to packages
