+++ packages/framework/src/Model/Product/Flag/FlagFacade.php
@@ -140,4 +140,33 @@ class FlagFacade
     {
         return $this->flagRepository->getFlagIdsByUuids($flagUuids);
     }
+
+    /**
+     * @param string $locale
+     * @return \Shopsys\FrameworkBundle\Model\Product\Flag\Flag[]
+     */
+    public function getAllVisibleFlags(string $locale): array
+    {
+        return $this->flagRepository->getAllVisibleFlags($locale);
+    }
+
+    /**
+     * @param string $uuid
+     * @param string $locale
+     * @return \Shopsys\FrameworkBundle\Model\Product\Flag\Flag
+     */
+    public function getVisibleByUuid(string $uuid, string $locale): Flag
+    {
+        return $this->flagRepository->getVisibleByUuid($uuid, $locale);
+    }
+
+    /**
+     * @param int $flagId
+     * @param string $locale
+     * @return \Shopsys\FrameworkBundle\Model\Product\Flag\Flag
+     */
+    public function getVisibleFlagById(int $flagId, string $locale): Flag
+    {
+        return $this->flagRepository->getVisibleFlagById($flagId, $locale);
+    }
 }
