--- project-base/app/config/graphql/types/Query/Flag/FlagQuery.types.yaml
@@ -1,17 +1,0 @@
-FlagQuery:
-    type: object
-    config:
-        fields:
-            flags:
-                type: '[Flag!]'
-                resolve: "@=query('flagsQuery')"
-                description: "Returns a complete list of the flags"
-            flag:
-                type: 'Flag'
-                resolve: "@=query('flagByUuidOrUrlSlugQuery', args['uuid'], args['urlSlug'])"
-                args:
-                    uuid:
-                        type: "Uuid"
-                    urlSlug:
-                        type: "String"
-                description: "Returns a flag by uuid or url slug"
