+++ project-base/app/config/graphql/types/Query/Query.types.yaml
@@ -6,7 +6,6 @@ Query:
         - 'CategoryQuery'
         - 'StoreQuery'
         - 'SliderQuery'
-        - 'FlagQuery'
         - 'OrderQuery'
         - 'PersonalDataQuery'
         - 'ProductQuery'
